Move helper functions to their own file
